ratbagd: add support for DBus signals
_RatbagdDBus now inherits from GObject so that we can use GObject's
signals when we receive a signal over DBus. This means that any users of
the bindings can just `Ratbagd*.connect("<signal>", <callback>)` to
be notified of any emitted signals.

Another advantage of inheriting to GObject is that the properties can
now be GObject properties. Again, this allows users of the bindings to
`Ratbagd*.connect("notify::<property>", <callback>)` to watch for
property changes.