Window: add a skeleton window
This is just a quick skeleton window for now to be able to test the
MouseMap widget. Eventually this will turn into the mockups.

piper.py is still around to steal code from; it will be removed
eventually when all useful bits have found a place in the new codebase.