Window: give perspectives a back button
I extended the perspective ``interface'' with another property called
`can_go_back`, which signifies if a perspective can go back into the
welcome perspective.

When this property is true, the window inserts a back button into the
perspective's titlebar that is visible when the perspective is presented
and there is more than one device connected.

Fixes #96.