ButtonDialog: don't crash on unknown mapping types
The ratbagd test device still uses a Key action type which we ignore in piper.
This leads to the correct "Unknown" display, but crashes when we click on the
button to configure it - self._mapping triggers and AttributeError.

Avoid this by simply resetting it to -1. At least that way we don't crash.