Fix icon rendering in the Qt SVG renderer
The Qt SVG Renderer only supports the SVG Tiny spec, so some features of SVG cannot be used without causing graphical glitches such as black backgrounds or certain effects simply not being rendered.

I've double checked that the appearance of the icons have not changed. The SVG code I removed wasn't for the visible parts of the icons. A side effect of this change is that the SVGs are a bit more optimized now.