LedDialog: fix buggy effect duration slider
Get the step increment from the Gtk.Adjustment instead of relying on a
magic number (500), which was wrong to begin with (actual value is 100).

Round the received float value to the nearest multiple of the step
increment, instead of rounding it down.

This fixes control of the slider through the arrow keys and improves the
accuracy of the "step snapping" when clicking the slider.

As a result, the UI now allows the user to pass values with a step (100)
instead of page (500) resolution to libratbag but that shouldn't be an
issue as drivers are expected to support any integer value within the
range 0-10000; see ratbag_led_set_effect_duration().